Switching things around and making new components within this view.
