+++ server/routes/feedback.router.js
@@ -105,20 +105,7 @@ router.get('/employee', (req, res) => {
 //will get all feedback count for specific employee
 router.get('/employeeFeedbackCount/1', (req, res) => {
     console.log('in GET /employeeFeedbackCount');
-    const empFeedbackCntQuery = `SELECT DISTINCT "employee"."first_name",
-	                            (SELECT COUNT ("feedback"."quality_id")
-	                            FROM "feedback" WHERE "feedback"."quality_id" = 1 AND "feedback"."employee_id" = 1) as Praise,
-	                            (SELECT COUNT ("feedback"."quality_id")
-	                            FROM "feedback" WHERE "feedback"."quality_id" = 2 AND "feedback"."employee_id" = 1) as Instruct,
-	                            (SELECT COUNT ("feedback"."quality_id")
-	                            FROM "feedback" WHERE "feedback"."quality_id" = 3 AND "feedback"."employee_id" = 1) as Correct
-                                FROM "employee"
-                                JOIN "feedback"
-                                ON "feedback"."employee_id" = "employee"."id"
-                                JOIN "quality_types"
-                                ON "feedback"."quality_id" = "quality_types"."id"
-                                WHERE "employee"."id" = 1
-                                GROUP BY "employee"."id", "employee"."first_name", "quality_types"."name";`;
+    const empFeedbackCntQuery = `SELECT * FROM "feedback" WHERE "employee_id" =1;`;
     pool.query(empFeedbackCntQuery)
         .then(result => res.send(result.rows))
         .catch(error => {
