+++ src/components/EmployeesView/IndividualEmployeeSummaryView/IndividualEmployeeSummary.css
@@ -1,21 +1,10 @@
-.btn {
-    margin: 0px;
-    top: auto;
-    right: 20px;
-    bottom: 20px;
-    position: fixed;
-    float: right;
-}
 
 .outer {
-    position: relative;
-    margin: 0 auto;
     display: flex;
     flex-direction: column;
 }
 
 .container {
-    overflow-y: auto;
     margin-right: 16px;
     padding: 4px;
 }
